feat: organise styles
