+++ packages/libs/react-ui/src/styles/index.ts
@@ -2,4 +2,4 @@ export { sprinkles, breakpoints } from './sprinkles.css';
 export type { Sprinkles } from './sprinkles.css';
 export { vars, darkThemeClass } from './vars.css';
 export type { ColorType } from './vars.css';
-export { colorPalette } from './colors';
+export { colorPalette } from './colors.css';
